multiple fix

+++ server/src/main/java/org/apache/iotdb/db/engine/version/SimpleFileVersionController.java
@@ -109,7 +109,9 @@ public class SimpleFileVersionController implements VersionController {
   private void persist() throws IOException {
     File oldFile = SystemFileFactory.INSTANCE.getFile(directoryPath, FILE_PREFIX + prevVersion);
     File newFile = SystemFileFactory.INSTANCE.getFile(directoryPath, FILE_PREFIX + currVersion);
-    FileUtils.moveFile(oldFile, newFile);
+    if (oldFile.exists()) {
+      FileUtils.moveFile(oldFile, newFile);
+    }
     logger.info("Version file updated, previous: {}, current: {}",
         oldFile.getAbsolutePath(), newFile.getAbsolutePath());
     prevVersion = currVersion;
